Fix download file action test
